aplicativo funcional

+++ app/src/main/java/com/example/checked/infra/CheckListPersistence.java
@@ -76,7 +76,7 @@ public class CheckListPersistence {
         return true;
     }
 
-    public ItemChecklist selectByID(int id){
+    public ItemChecklist selectByID(long id){
         ItemChecklist item = null;
         Cursor cursor;
